Code and data corrections.

+++ Strat-O-Matic-SqLite/Strat-O-Matic-SqLiteDoc.h
@@ -77,7 +77,7 @@ public:
 	int GetBatterID(CStringA strFirstName, CStringA strLastName);
 	int GetPitcherID(CStringA strFirstName, CStringA strLastName);
 	int PitcherInsert(CStringA strFirstName, CStringA strLastName, float OBChanceHomeRun, float OBChanceTriple, float OBChanceDouble, float OBChanceSingle, float OBChanceWalk, float ChanceDoublePlay, float OBChanceHomeRunRight, float OBChanceTripleRight, float OBChanceDoubleRight, float OBChanceSingleRight, float OBChanceWalkRight, float ChanceDoublePlayRight, float OBChanceHomeRunLeft, float OBChanceTripleLeft, float OBChanceDoubleLeft, float OBChanceSingleLeft, float OBChanceWalkLeft, float ChanceDoublePlayLeft, float OBChanceBasic, float OBChanceLeft, float OBChanceRight, int Starter, int Relief, int Throws, int Bunting, int Hold, int WP, int Balk, int Pitcher, int ER1, int TeamID);
-	int PitcherStatsInsert(int Wins, int Loss, int Saves, int InningsPitched, int ER, int Hits, int Walks, int Strikeouts, int HomeRuns, int Games, int CompleteGames, int Starts, float ERA, float WHIP, int PitcherID, int TeamID);
+	int PitcherStatsInsert(int Wins, int Loss, int Saves, float InningsPitched, int ER, int Hits, int Walks, int Strikeouts, int HomeRuns, int Games, int CompleteGames, int Starts, float ERA, float WHIP, int PitcherID, int TeamID);
 	afx_msg void OnSqlInsertPitcher();
 	afx_msg void OnSqlInsertPitcherStats();
 	int GetBatterID(CStringA strFirstName, CStringA strLastName, int TeamID);
